- moved helpers from decodeDataField into utils/transactions/getDecodedDataField - skip smart and decimal decoding on datafields with a length larger than 36000 characters as this would often slow down/break broswers
