+++ src/constants/transactions.constants.ts
@@ -10,3 +10,4 @@ export const SAME_SHARD_MAX_TOAST_WIDTH_PERCENT_DECREASE = 75;
 export const CROSS_SHARD_MAX_TOAST_WIDTH_PERCENT_DECREASE = 25;
 export const PROGRESS_INTERVAL_DURATION_MS = 100;
 export const DEFAULT_TOAST_LIEFTIME = 5000;
+export const MAX_DECODE_TX_DATA_LENGTH = 36000;
