+++ src/utils/transactions/index.ts
@@ -1,4 +1,5 @@
 export * from './explorerUrlBuilder';
 export * from './getActiveTransactionsStatus';
+export * from './getDecodedDataField';
 export * from './getHumanReadableTimeFormat';
 export * from './getTransactionsHistory';
